Viser read-only view av meldekort-behandlingen dersom innlogget bruker ikke er saksbehandler for det aktuelle meldekortet
